1、修改makefile支持设置编译标志和连接标志。 2、把编译器检测级别设置为最高。 3、修正几个编译器检查出来的，函数漏写返回值的错误。 4、实现FrameBuffer类中像素点的获取与写入操作。

+++ video/Bitmap.cc
@@ -25,7 +25,7 @@ int Bitmap::loadFromBmpFile(const char *filename)
 
   std::tr1::shared_ptr<FILE> filePtr(fp, ::fclose);
 
-  int ret = ::fread(&fileHeader, 1, sizeof(BitmapFileHeader), fp);
+  size_t ret = ::fread(&fileHeader, 1, sizeof(BitmapFileHeader), fp);
   if (ret != sizeof(BitmapFileHeader) || fileHeader.type != kBmpFileMagic)
   {
     return -1; 
@@ -75,7 +75,7 @@ int Bitmap::storeToBmpFile(const char *filename)
 
   std::tr1::shared_ptr<FILE> filePtr(fp, ::fclose);
 
-  int ret = ::fwrite(&fileHeader, 1, sizeof(BitmapFileHeader), fp);
+  size_t ret = ::fwrite(&fileHeader, 1, sizeof(BitmapFileHeader), fp);
   if (ret != sizeof(BitmapFileHeader))
   {
     return -1;
